+++ src/App.jsx
@@ -1,6 +1,5 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import { Home } from "./components/Home";
-import { Login } from "./components/Login";
+import { Home, Login } from "./views";
 
 function App() {
   return (
