+++ src/assets/styles/index.css
@@ -4,3 +4,10 @@ body,
   min-height: 100%;
   height: 100%;
 }
+.main {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 24px;
+  height: 100%;
+}
