+++ src/layouts/index.js
@@ -1,0 +1,1 @@
+export { default as DefaultLayout } from "./DefaultLayout";
