+++ src/main.jsx
@@ -1,11 +1,11 @@
 import React from "react";
-import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
+import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import store from "./redux/store";
 
 import CssBaseline from "@mui/material/CssBaseline";
-import { ThemeProvider } from "@mui/material";
+import { ThemeProvider, StyledEngineProvider } from "@mui/material";
 import theme from "./utils/theme/ThemeConfig";
 import "./assets/styles/index.css";
 
@@ -13,8 +13,10 @@ ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     <Provider store={store}>
       <ThemeProvider theme={theme("light")}>
-        <CssBaseline />
-        <App />
+        <StyledEngineProvider injectFirst>
+          <CssBaseline />
+          <App />
+        </StyledEngineProvider>
       </ThemeProvider>
     </Provider>
   </React.StrictMode>
